+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Add `dir` argument to `filter` function
   * Support using tokens multiple times
 
 1.3.1 / 2014-10-01